Voting-Authority
- add GITHUB env variables
- add GITHUB env variables to Dockerfile as build args
- adjust docker-compose.yml and shell scripts accordingly

+++ voting-authority/docker-start.sh
@@ -18,30 +18,21 @@ echo "The mode is: $mode"
 # - the config file with all IPs and ports
 ###########################################
 globalConfig=$parentDir/system.json
+githubConfig=$parentDir/github.json
+
+# check if the github config json exists -> if not we stop the process
+if [ ! -f "$githubConfig" ]; then
+    echo "$githubConfig doesn't exist! Please create it!"
+    exit
+fi
 
 ###########################################
 # Cleanup
 ###########################################
-rm -rf $dir/backend/mp-crypto
 rm -f $dir/backend/.env
 rm -f $dir/frontend/.env
 rm -f $dir/.env
-
-###########################################
-# Unlink mp-crypto as we don't want the 
-# symbolic link inside the container
-# the build will fail otherwise
-###########################################
-sudo npm unlink --no-save mp-crypto
 rm -rf $dir/backend/node_modules
-
-########################################
-# mp-crypto library
-#######################################
-cryptoPath=$parentDir/crypto
-cp -r $cryptoPath $dir/backend/mp-crypto
-rm -rf $dir/backend/mp-crypto/node_modules
-rm -rf $dir/backend/mp-crypto/dist
 
 ###########################################
 # ENV variables
@@ -60,6 +51,11 @@ PARITY_NODE_PORT=$(cat $globalConfig | jq .services.sealer_parity_1.port)
 PARITY_NODE_IP=$(cat $globalConfig | jq .services.sealer_parity_1.ip.$mode | tr -d \")
 # - Specify NODE_ENV
 NODE_ENV=$mode
+# - Specify the Github credentials
+GITHUB_TOKEN=$(cat $githubConfig | jq .github.token | tr -d \")
+GITHUB_EMAIL=$(cat $githubConfig | jq .github.email | tr -d \")
+GITHUB_USER=$(cat $githubConfig | jq .github.user | tr -d \")
+
 
 ###########################################
 # write ENV variables into .env
@@ -71,6 +67,9 @@ echo VOTING_AUTH_FRONTEND_IP=$VOTING_AUTH_FRONTEND_IP >> $dir/.env
 echo PARITY_NODE_PORT=$PARITY_NODE_PORT >> $dir/.env
 echo PARITY_NODE_IP=$PARITY_NODE_IP >> $dir/.env
 echo NODE_ENV=$NODE_ENV >> $dir/.env
+echo GITHUB_TOKEN=$GITHUB_TOKEN >> $dir/.env
+echo GITHUB_USER=$GITHUB_USER >> $dir/.env
+echo GITHUB_EMAIL=$GITHUB_EMAIL >> $dir/.env
 
 ###########################################
 # docker network
@@ -84,5 +83,5 @@ $parentDir/docker-network.sh $network_name
 # start containers
 ###########################################
 cd $dir
-docker-compose -p vote-auth -f docker-compose.yml up  --detach
+docker-compose -p vote-auth -f docker-compose.yml up --build --detach
 rm -f $dir/.env
